+++ tests/test_communication_channel.py
@@ -102,10 +102,9 @@ class TestCommunicationChannel(unittest.TestCase):
 
     # update_preferences_by_category()
     def test_update_preferences_by_category(self, m):
-        register_uris(
-            {
-                'communication_channel': ['update_preferences_by_category']
-            }, m)
+        register_uris({
+            'communication_channel': ['update_preferences_by_category']
+        }, m)
         category = 'course_content'
         frequency = 'daily'
 
@@ -159,3 +158,16 @@ class TestCommunicationChannel(unittest.TestCase):
                 notification_preferences=no_frequency_notification_preferences
             )
         )
+
+    # delete()
+    def test_delete(self, m):
+        register_uris({
+            'communication_channel': ['create_comm_channel', 'delete_comm_channel']
+        }, m)
+
+        channel = {
+            "type": "email",
+            "address": "username@example.org"
+        }
+        new_channel = self.user.create_communication_channel(communication_channel=channel)
+        self.assertTrue(new_channel.delete())